+++ lib/scala/filewatcher/src/main/java/org/enso/filewatcher/package-info.java
@@ -1,0 +1,1 @@
+package org.enso.filewatcher;
